# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * BougieBams API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ApiProduct {
  id: string;
  sku: string;
  name: string;
  description: string;
  price: number;
  category: string;
  inStock: boolean;
  /** @nullable */
  imagePath?: string | null;
}

export interface ProductInput {
  sku: string;
  name: string;
  description: string;
  price: number;
  category: string;
  inStock: boolean;
}

export interface ProductUpdate {
  name?: string;
  description?: string;
  price?: number;
  category?: string;
  inStock?: boolean;
}

export interface ProductListResponse {
  products: ApiProduct[];
}

export interface ProductResponse {
  product: ApiProduct;
}

export type ApiEventCategory = typeof ApiEventCategory[keyof typeof ApiEventCategory];


export const ApiEventCategory = {
  'In-Person': 'In-Person',
  Virtual: 'Virtual',
  Tournament: 'Tournament',
  Workshop: 'Workshop',
} as const;

export interface ApiEvent {
  id: number;
  title: string;
  description: string;
  date: string;
  time: string;
  location: string;
  /** @nullable */
  priceCents?: number | null;
  category: ApiEventCategory;
  /** @nullable */
  imagePath?: string | null;
  totalSpots: number;
  spotsLeft: number;
  host: string;
  published: boolean;
}

export type EventInputCategory = typeof EventInputCategory[keyof typeof EventInputCategory];


export const EventInputCategory = {
  'In-Person': 'In-Person',
  Virtual: 'Virtual',
  Tournament: 'Tournament',
  Workshop: 'Workshop',
} as const;

export interface EventInput {
  title: string;
  description: string;
  date: string;
  time: string;
  location: string;
  /** @nullable */
  priceCents?: number | null;
  category: EventInputCategory;
  totalSpots: number;
  spotsLeft: number;
  host: string;
  published?: boolean;
}

export type EventUpdateCategory = typeof EventUpdateCategory[keyof typeof EventUpdateCategory];


export const EventUpdateCategory = {
  'In-Person': 'In-Person',
  Virtual: 'Virtual',
  Tournament: 'Tournament',
  Workshop: 'Workshop',
} as const;

export interface EventUpdate {
  title?: string;
  description?: string;
  date?: string;
  time?: string;
  location?: string;
  /** @nullable */
  priceCents?: number | null;
  category?: EventUpdateCategory;
  totalSpots?: number;
  spotsLeft?: number;
  host?: string;
  published?: boolean;
}

export interface EventListResponse {
  events: ApiEvent[];
}

export interface EventResponse {
  event: ApiEvent;
}

export interface EventStats {
  total: number;
  published: number;
  upcoming: number;
  totalRegistrations: number;
}

export interface RegistrationInput {
  eventId: number;
  name: string;
  email: string;
  notes?: string;
}

export interface RegistrationRecord {
  id: number;
  eventId: number;
  name: string;
  email: string;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}

export interface RegistrationResult {
  registration: RegistrationRecord;
}

export interface ApiBlogPost {
  id: number;
  slug: string;
  title: string;
  excerpt: string;
  content: string;
  category: string;
  author: string;
  /** @nullable */
  imagePath?: string | null;
  published: boolean;
  createdAt: string;
  updatedAt: string;
}

export interface BlogPostInput {
  slug: string;
  title: string;
  excerpt: string;
  content: string;
  category: string;
  author: string;
  published?: boolean;
}

export interface BlogPostUpdate {
  title?: string;
  excerpt?: string;
  content?: string;
  category?: string;
  author?: string;
  published?: boolean;
}

export interface BlogPostListResponse {
  posts: ApiBlogPost[];
}

export interface BlogPostResponse {
  post: ApiBlogPost;
}

export interface SubscribeInput {
  email: string;
  source?: string;
  discountCode?: string;
}

export interface SubscribeResult {
  message: string;
  /** @nullable */
  discountCode?: string | null;
}

export interface Subscriber {
  id: number;
  email: string;
  /** @nullable */
  source?: string | null;
  /** @nullable */
  discountCode?: string | null;
  createdAt: string;
}

export interface SubscriberListResponse {
  subscribers: Subscriber[];
}

export interface ContactInput {
  name: string;
  email: string;
  subject: string;
  message: string;
}

export interface AdminLoginInput {
  password: string;
}

export interface AdminAuthToken {
  token: string;
}

export interface AdminStats {
  totalSubscribers: number;
  totalEvents: number;
  totalProducts: number;
  totalBlogPosts: number;
}

export interface MessageResponse {
  message: string;
}