# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: BougieBams API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /events:
    get:
      operationId: listEvents
      tags: [events]
      summary: List published events
      responses:
        "200":
          description: Events list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventListResponse"
    post:
      operationId: createEvent
      tags: [events]
      summary: Create an event (admin)
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventInput"
      responses:
        "201":
          description: Created event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventResponse"
        "401":
          description: Unauthorized
  /events/stats:
    get:
      operationId: getEventStats
      tags: [events]
      summary: Event stats summary (admin)
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Event stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventStats"
  /events/{id}:
    get:
      operationId: getEvent
      tags: [events]
      summary: Get a single event
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventResponse"
        "404":
          description: Not found
    patch:
      operationId: updateEvent
      tags: [events]
      summary: Update an event (admin)
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventUpdate"
      responses:
        "200":
          description: Updated event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventResponse"
        "401":
          description: Unauthorized
    delete:
      operationId: deleteEvent
      tags: [events]
      summary: Delete an event (admin)
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "401":
          description: Unauthorized
...
  /blog:
    get:
      operationId: listBlogPosts
      tags: [blog]
      summary: List published blog posts
      responses:
        "200":
          description: Blog posts list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlogPostListResponse"
    post:
      operationId: createBlogPost
      tags: [blog]
      summary: Create a blog post (admin)
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BlogPostInput"
      responses:
        "201":
          description: Created post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlogPostResponse"
        "401":
          description: Unauthorized
  /blog/{slug}:
    get:
      operationId: getBlogPost
      tags: [blog]
      summary: Get a blog post by slug
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Blog post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlogPostResponse"
        "404":
          description: Not found
    patch:
      operationId: updateBlogPost
      tags: [blog]
      summary: Update a blog post (admin)
      security:
        - BearerAuth: []
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BlogPostUpdate"
      responses:
        "200":
          description: Updated post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlogPostResponse"
        "401":
          description: Unauthorized
    delete:
      operationId: deleteBlogPost
      tags: [blog]
      summary: Delete a blog post (admin)
      security:
        - BearerAuth: []
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
        "401":
          description: Unauthorized

  # ── Subscribers ──────────────────────────────────────────────────────────
...
  /admin/auth/login:
    post:
      operationId: adminLogin
      tags: [auth]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminLoginInput"
      responses:
        "200":
          description: Login success with token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminAuthToken"
        "401":
          description: Invalid password

  # ── Admin stats ──────────────────────────────────────────────────────────
  /admin/stats:
    get:
      operationId: getAdminStats
      tags: [admin]
      summary: Dashboard summary stats (admin)
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Admin dashboard stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminStats"
        "401":
          description: Unauthorized
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    ApiEvent:
      type: object
      required: [id, title, description, date, time, location, category, totalSpots, spotsLeft, host, published]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        date:
          type: string
        time:
          type: string
        location:
          type: string
        priceCents:
          type: ["integer", "null"]
        category:
          type: string
          enum: ["In-Person", "Virtual", "Tournament", "Workshop"]
        imagePath:
          type: ["string", "null"]
        totalSpots:
          type: integer
        spotsLeft:
          type: integer
        host:
          type: string
        published:
          type: boolean
    EventInput:
      type: object
      required: [title, description, date, time, location, category, totalSpots, spotsLeft, host]
      properties:
        title:
          type: string
        description:
          type: string
        date:
          type: string
        time:
          type: string
        location:
          type: string
        priceCents:
          type: ["integer", "null"]
        category:
          type: string
          enum: ["In-Person", "Virtual", "Tournament", "Workshop"]
        totalSpots:
          type: integer
        spotsLeft:
          type: integer
        host:
          type: string
        published:
          type: boolean
    EventUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        date:
          type: string
        time:
          type: string
        location:
          type: string
        priceCents:
          type: ["integer", "null"]
        category:
          type: string
          enum: ["In-Person", "Virtual", "Tournament", "Workshop"]
        totalSpots:
          type: integer
        spotsLeft:
          type: integer
        host:
          type: string
        published:
          type: boolean
    EventListResponse:
      type: object
      required: [events]
      properties:
        events:
          type: array
          items:
            $ref: "#/components/schemas/ApiEvent"
    EventResponse:
      type: object
      required: [event]
      properties:
        event:
          $ref: "#/components/schemas/ApiEvent"
    EventStats:
      type: object
      required: [total, published, upcoming, totalRegistrations]
      properties:
        total:
          type: integer
        published:
          type: integer
        upcoming:
          type: integer
        totalRegistrations:
          type: integer

    # ── Registration schemas ─────────────────────────────────────────────
...
    ApiBlogPost:
      type: object
      required: [id, slug, title, excerpt, content, category, author, published, createdAt, updatedAt]
      properties:
        id:
          type: integer
        slug:
          type: string
        title:
          type: string
        excerpt:
          type: string
        content:
          type: string
        category:
          type: string
        author:
          type: string
        imagePath:
          type: ["string", "null"]
        published:
          type: boolean
        createdAt:
          type: string
        updatedAt:
          type: string
    BlogPostInput:
      type: object
      required: [slug, title, excerpt, content, category, author]
      properties:
        slug:
          type: string
        title:
          type: string
        excerpt:
          type: string
        content:
          type: string
        category:
          type: string
        author:
          type: string
        published:
          type: boolean
    BlogPostUpdate:
      type: object
      properties:
        title:
          type: string
        excerpt:
          type: string
        content:
          type: string
        category:
          type: string
        author:
          type: string
        published:
          type: boolean
    BlogPostListResponse:
      type: object
      required: [posts]
      properties:
        posts:
          type: array
          items:
            $ref: "#/components/schemas/ApiBlogPost"
    BlogPostResponse:
      type: object
      required: [post]
      properties:
        post:
          $ref: "#/components/schemas/ApiBlogPost"

    # ── Subscriber schemas ───────────────────────────────────────────────
...
    AdminLoginInput:
      type: object
      required: [password]
      properties:
        password:
          type: string
    AdminAuthToken:
      type: object
      required: [token]
      properties:
        token:
          type: string

    # ── Admin stats ──────────────────────────────────────────────────────
    AdminStats:
      type: object
      required: [totalSubscribers, totalEvents, totalProducts, totalBlogPosts]
      properties:
        totalSubscribers:
          type: integer
        totalEvents:
          type: integer
        totalProducts:
          type: integer
        totalBlogPosts:
          type: integer

    # ── Shared ───────────────────────────────────────────────────────────
...